requirements.txt
torch
numpy 
torchvision 
transformers
pillow
albumentations
pandas 
monai
tqdm
torchmetrics